requirements.txt
python-telegram-bot
python-dotenv
sqlalchemy
pillow